ai-modules/modules/rehabilitation/requirements-prod.txt
# Production Requirements for Rehabilitation AI Module
# Generated: December 2025
# Python: 3.8+

# Core FastAPI Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6

# HTTP client
httpx==0.25.2

# Environment management
python-dotenv==1.0.0

# Machine Learning - Core
numpy==1.26.2
pandas==2.1.4
scipy==1.11.4
scikit-learn==1.3.2

# XGBoost
xgboost==2.0.2

# Model serialization
joblib==1.3.2

# Natural Language Processing - Deep Learning
torch==2.1.0
transformers==4.35.0
sentence-transformers==2.2.2
nltk==3.8.1

# Authentication
PyJWT==2.8.1

# Optional: For production deployment
# gunicorn==21.2.0  # WSGI server
# python-multipart==0.0.6  # Already included
# python-jose==3.3.0  # For advanced JWT handling

# Optional: For monitoring and logging
# python-json-logger==2.0.7  # JSON formatted logging
# prometheus-client==0.18.0  # Prometheus metrics
